* gwork:
  Exit script on error. Add dummy-run.
  Notes of software renderer as default options.
  Update changelog for Irrlicht.
  Bump copyright year.
  The DebugBreak header already contains this information so no need to duplicate.
  fixed RGB octets showing as characters
  Fixing up Irrlicht cmake
  Proper naming
  Fixed issues found by billy
  Irrlicht3D Renderer
  Irrlicht Renderer
  Update Irrlicht Renderer
  Fix typos.
  Some introduction documentation.
  Irrlicht Renderer + Input Handler

# Conflicts:
#	docs/updocs.sh
#	source/platform/platforms/WindowsPlatform.cpp

+++ docs/updocs.sh
@@ -1,6 +1,8 @@
 #!/usr/bin/env bash
 # Update online documentation
 
+set -e
+
 PROJDIR=tmp-proj
 DOCDIR=tmp-docs
 DOCOUT=$PROJDIR/build/doc/html
@@ -13,17 +15,18 @@ git clone https://github.com/billyquith/gwork.git $DOCDIR -b gh-pages --depth=1
 
 pushd $PROJDIR
 mkdir build && cd build
-cmake .. -G Ninja -DWANT_RENDERER_NULL=ON
-ninja doc
+cmake .. -G Ninja -DRENDER_NULL=ON
+cmake --build . --target doc
 cd ..
 popd
 
-rm -r $DOCDIR/*
+[ -d $DOCDIR] && rm -r $DOCDIR/*
 cp -r $DOCOUT/* $DOCDIR
 
 pushd $DOCDIR
 git add --all --verbose .
 git status
+if [[ "$1" == "-d" ]]; then exit 0; fi
 read -p "Commit message: " MESSAGE
 git commit -m "${MESSAGE:-"Update docs"}"
 git push
